Huuuge Progress!
I figured out how to make it so when you press the Circle Switch it opens the Circle Gate and when you press the Triangle Switch, it opens the Triangle Gate!  I moved on next to the Hexagon Gate & Switch, but I ran into a problem.  I've been working at this for like 2 hours now, so I'm going to take a break.