Additions and changes (Mark 2)
Added cameras in place for the title screen level for the Credits and The Pieces area (camera change needs to be added). Altered the Move counter to have more discriptive text, need to fully understand the dimensions on the FOV area. Alligned the camera switch buttons along with the title screen items for more pleasing visual looks.